Capture multiple screenshots (#5)
Each click on the window captures a new screenshot of the middle of the screen.

We store all captured screenshots in the same `pixmap` in a film strip layout stacked on top of each other. We only keep track of the N most recent screenshots (defined by max_screenshots_shown).