Track any click (#6)
Before this PR, we could only detect when you clicked on our own window.

With this PR, we can detect a click event wherever it happens to be and trigger a screenshot. This way when you're playing game, screenshots are still taken as you play.

Using the X Input Extension to track raw clicks regardless of whether our window has focus or the click occurs on our window.